#5 refactor SlackNameHandlerService - read markers for parced Uuid from property file. And refactor tests for this class.
