+++ application.properties
@@ -5,3 +5,5 @@ endpoint.interview=/achieve/interview
 endpoint.userSearch=/users/search
 gamification.baseURL=http://gamification.juja.com.ua
 user.baseURL=http://user.juja.com.ua
+parcedUuid.startMarker=@#
+parcedUuid.finishMarker=#@
